로직들은 함수화 및 경로 분리
